charaobj: widen statAttack/calcCastTime se+item values to int
permute_fn surfaced these; corrected so the underlying loads stay unsigned
(lhz, matching the target) while the working values are int:
- statAttack: seSpec read unsigned, held as int for the (>>8)*1000 decode.
- calcCastTime: itemNo read unsigned, held as int for the == 0x1F8 checks.

statAttack 84.12%->86.49%; calcCastTime 88.19%->88.65%; unit 81.12%->81.21%.

+++ src/charaobj.cpp
@@ -2964,7 +2964,7 @@ void CGCharaObj::statAttack()
 	unsigned char* itemData = reinterpret_cast<unsigned char*>(Game.unkCFlatData0[2]) + m_itemId * 0x48;
 	unsigned short seFrame = *reinterpret_cast<unsigned short*>(itemData + 0x3A);
 	if ((seFrame & 0x8000) == 0 && m_stateFrame == seFrame) {
-		unsigned int seSpec = *reinterpret_cast<unsigned short*>(itemData + 0x38);
+		int seSpec = *reinterpret_cast<unsigned short*>(itemData + 0x38);
 		if (seSpec != 0) {
 			int seNo = (seSpec == 0xFFFF) ? 0 : ((seSpec & 0xFF) + ((seSpec >> 8) * 1000));
 			playSe3D(seNo, 0x32, 0x96, 0, 0);
@@ -3123,7 +3123,7 @@ int CGCharaObj::calcCastTime(int itemId)
 		castScale = kOneF32;
 	}
 
-	unsigned short itemNo = *reinterpret_cast<unsigned short*>(itemData + 0x0);
+	int itemNo = *reinterpret_cast<unsigned short*>(itemData + 0x0);
 	unsigned short itemType = *reinterpret_cast<unsigned short*>(itemData + 0xE);
 
 	if (itemNo != 0x1F8 && itemType == 2) {
